Updated Index.qmd
Incorporated Adrian's feedback. Removed the Linear Model, made sure the first scatterplot had the size of the dots as the number of reviews a movie received. His second point of feedback asked us to do what we had already done with the first scatterplot, which was include a line of best fit with method = 'lm'.